requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
av==8.0.2
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==3.1.5
bokeh==2.2.1
bz2file==0.98
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dlib==19.24.2
docopt==0.6.2
entrypoints==0.3
facenet-pytorch==2.5.1
ffmpeg-python==0.2.0
filelock==3.3.1
future==0.18.2
gast==0.3.3
gdown==4.1.1
google-pasta==0.2.0
grpcio==1.30.0
#gulpio==540.66
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
livelossplot==0.5.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
#ngt==1.14.6
notebook==6.0.3
numpy==1.19.1
opencv-python==4.1.2.30
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
Pillow-SIMD==7.0.0.post3
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pybind11==2.10.0
pyDeprecate==0.3.2
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.25.0
scikit-image==0.16.2
scikit-learn==0.24.2
scipy==1.2.1
seaborn==0.11.2
Send2Trash==1.5.0
sh==1.14.1
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thop==0.1.1.post2209072238
threadpoolctl==3.0.0
torch==1.10.2
torch-lr-finder==0.2.1
torchmetrics==0.8.2
torchsummary==1.5.1
torchvideo==0.0.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
